pyproject.toml
[tool.poetry]
name = "video-stabilizer"
version = "1.0.0"
description = "A video stabilization tool using Python and OpenCV"
authors = ["Dein Name <deine.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
opencv-python = "^4.5.3"
numpy = "^1.21.0"
tqdm = "^4.61.2"
pyinstaller = "^5.5"
pygments = "^2.9.0"
pillow = "^11.1.0"

[tool.poetry.scripts]
video-stabilizer = "video_stabilizer:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
